Dispatch redux-form actions to indicate that saving is complete
We can use `redux-form` `startSubmit` and `stopSubmit` functions to
toggle the value of the `submitting` prop. Both functions take the name
of the form that is being submitted - https://goo.gl/PGVWSW

Note that because of a bug in `redux-form`, `startSubmit` maps to
`startSave` and `stopSubmit` maps to `stopSave` - https://goo.gl/PUiLMv.